+++ CHANGELOG.md
@@ -1,1 +1,4 @@
+## [binjr-adapter-demo 1.0.0](https://github.com/binjr/binjr/releases/tag/1.0.0)
+Released on Thu, 17 Oct 2019
 
+ * Initial release